requirement.txt
allure-pytest==2.9.45
pytest==8.0.0
selenium==4.20.0
requests==2.20.0
pytest-assume==2.2.0